+++ src/main/java/at/hannibal2/skyhanni/utils/renderables/container/RenderableInventory.kt
@@ -98,7 +98,7 @@ object RenderableInventory {
                 val uvArray = uv.getUvCoords()
                 drawInsideFixedSizedImage(
                     if (uv == SlotsUv.CENTER)
-                        items[index++]?.let { item(it, scale, 0, 0, false) } ?: emptySlot
+                        items[index++]?.let { item(it, scale.toFloat(), 0, 0, false) } ?: emptySlot
                     else Renderable.empty(),
                     inventoryTextures,
                     (uv.width() * scale).toInt(),
